refactor: extract processed Hoard from config module
