+++ src/checkers/history/last_paths.rs
@@ -5,7 +5,7 @@
 //! explanation of why this is useful.
 
 use super::super::Checker;
-use crate::config::hoard::Hoard;
+use crate::hoard::Hoard;
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
